Добавлено новое свойство drain-backlog-on-startup в YandexBotProperties для управления сбросом накопленной очереди при старте бота. Обновлены классы YandexMessengerClient, YandexPollingOffsetStore и YandexPollingWorker для поддержки работы с long типом offset и улучшения обработки обновлений. Изменения способствуют более надежной работе бота и предотвращают повторную обработку устаревших обновлений.

+++ assistant-yandexbot/src/main/java/ru/georgdeveloper/assistantyandexbot/client/YandexMessengerClient.java
@@ -107,7 +107,7 @@ public class YandexMessengerClient {
 		}
 	}
 
-	public JsonNode getUpdates(int limit, int offset) throws Exception {
+	public JsonNode getUpdates(int limit, long offset) throws Exception {
 		// Polling режим: получаем следующую пачку обновлений.
 		Map<String, Object> body = new LinkedHashMap<>();
 		body.put("limit", limit);
